refactor: refactor and cleanup
